refactor extractor_preprocessing and feature extractor wrapper implementation
