+++ FE_configuration.py
@@ -1,264 +1,6 @@
-import json
-import shutil
-from sklearn.pipeline import Pipeline
-import tsfel
-from tsfresh.feature_extraction import  EfficientFCParameters
-
-from transformers.feature_extractors import TSFELLagFeatureExtractor, TSFreshLagFeatureExtractor
-from transformers.feature_selection import FeatureSelectionWrapper
-from transformers.imputation import ImputationWrapper
-from utils import clean_column_names, restore_column_names, train_test_split_by_time
-from lightgbm import LGBMClassifier
-import numpy as np
-
-def train_model(X_train, y_train, model, protected_cols=None):
-    X_train = X_train.copy()
-    
-    X_train['DateTime'] = X_train['DateTime'].astype('int64')
-    
-    X_train.drop(columns=protected_cols, inplace=True, errors='ignore')
-
-    model.fit(X_train, y_train)
-    return model
-
-
-def get_top_features(model, X_train, n_features=None, protected_cols=None):
-    feature_importances = model.feature_importances_
-    
-    feature_names = [col for col in X_train.columns if col not in protected_cols]
-
-    # If n_features is None, use all features
-    n_features = len(feature_names) if n_features is None else n_features
-    top_features = sorted(zip(feature_importances, feature_names), reverse=True)[:n_features]
-    top_feature_names = [feature[1] for feature in top_features]
-    return top_feature_names
-
-# function to select from a list, the number elements that have '_0_' in its name
-def select_tsfel_features(features, n_features=10):
-    """
-    Selects features that have'_0_' in their name (ones generated by tsfel).
-    """
-    tsfel_features = [s for s in features if '_0_' in s]
-
-    #for each tsfel features split by _0_ and keep the second part
-    tsfel_feature_names = [s.split('_0_')[1] for s in tsfel_features]
-
-    tsfel_features = []
-    for feature_name in tsfel_feature_names:
-        if feature_name not in tsfel_features:
-            tsfel_features.append(feature_name)
-        if len(tsfel_features) == n_features:
-            break
-    return tsfel_features
-
-def select_tsfresh_features(features, n_features=10):
-    tsfresh_features = [s for s in features if '__' in s]
-
-    tsfresh_feature_names = ['__'.join(s.split('__')[1:]) for s in tsfresh_features]
-
-    tsfresh_features = []
-    for feature_name in tsfresh_feature_names:
-        if feature_name not in tsfresh_features:
-            tsfresh_features.append(feature_name)
-        if len(tsfresh_features) == n_features:
-            break
-    return tsfresh_features
-
-def generate_tsfel_config(selected_features=None):
-    """
-    Creates a modified copy of the TSFEL configuration file.
-    
-    Args:
-        selected_features (list): List of feature names to keep enabled in config
-    
-    This function:
-    1. Reads the original TSFEL features.json
-    2. Creates a new config with only selected features enabled
-    3. Saves it to tsfel_config.json
-    """
-    config_path = tsfel.__path__[0] + "/feature_extraction/features.json"
-    
-    try:
-        # Read original config
-        with open(config_path, 'r') as file:
-            tsfel_config = json.load(file)
-
-        # Create modified configuration
-        modified_config = {}
-        for domain in tsfel_config:
-            modified_config[domain] = {}
-            for feature in tsfel_config[domain]:
-                modified_config[domain][feature] = tsfel_config[domain][feature].copy()
-                # Disable all features by default
-                modified_config[domain][feature]["use"] = "no"
-                # Enable only selected features
-                if selected_features and feature in selected_features:
-                    modified_config[domain][feature]["use"] = "yes"
-
-        # Save modified configuration to new file
-        CONFIG_FILE_LOCATION = 'tsfel_config.json'
-        with open(CONFIG_FILE_LOCATION, 'w') as outfile:
-            json.dump(modified_config, outfile, indent=4)
-            
-        print("Successfully created modified tsfel_config.json")
-        
-    except FileNotFoundError:
-        print(f"Error: Could not find original config file at {config_path}")
-    except json.JSONDecodeError:
-        print("Error: Invalid JSON in configuration file")
-    except Exception as e:
-        print(f"Error occurred: {str(e)}")
-
-    return CONFIG_FILE_LOCATION
-
-def get_top_dataset_features(data, COLUMN_CONFIG, n_features):
-    """Extract top features from the original dataset using LightGBM."""
-    # Split data and prepare for training
-    train_df, _ = train_test_split_by_time(data, train_ratio=0.7)
-    X_train, y_train = train_df.drop(columns=COLUMN_CONFIG['target_col']), train_df[COLUMN_CONFIG['target_col']]
-
-    # Train model and get top features
-    clf = LGBMClassifier()
-    model = train_model(X_train, y_train, clf, COLUMN_CONFIG['protected_cols'])
-    top_feature_names = get_top_features(model, X_train, n_features=n_features, protected_cols=COLUMN_CONFIG['protected_cols'])
-    print(f"Top {n_features} features: {top_feature_names}")
-    return top_feature_names, train_df
-
-def get_subset_features_df(train_df, top_feature_names, COLUMN_CONFIG):
-    """Extract subset of features from the original dataset based on top features."""
-    subset_features_df = train_df[top_feature_names + COLUMN_CONFIG['protected_cols']]
-    subset_features_df = subset_features_df.loc[:, ~subset_features_df.columns.duplicated()]
-
-    return subset_features_df
-
-def get_custumed_fc_parameters(final_tsfresh_features):
-    fc_parameters = {}
-
-    def convert_value(value):
-        """Helper function to convert string values to appropriate types"""
-        # Remove quotes if they exist
-        if value.startswith('"') and value.endswith('"'):
-            return value[1:-1]
-        
-        # Try converting to tuple if it contains parentheses
-        if '(' in value and ')' in value:
-            try:
-                # Extract numbers from (2, 5, 10, 20) format
-                nums = value.strip('()').split(',')
-                return tuple(float(n.strip()) for n in nums)
-            except:
-                return value
-                
-        # Try converting to float/int
-        try:
-            # Convert to float first
-            float_val = float(value)
-            # If it's a whole number, convert to int
-            if float_val.is_integer():
-                return int(float_val)
-            return float_val
-        except ValueError:
-            # Handle boolean values
-            if value.lower() == 'true':
-                return True
-            if value.lower() == 'false':
-                return False
-            # If all else fails, return original string
-            return value
-
-    for feature_full_name in final_tsfresh_features:
-        feature_name_parts = feature_full_name.split('__')
-        feature_name = feature_name_parts[0]
-        feature_attrs = feature_name_parts[1:]
-
-        # If no attributes, set to None
-        if not feature_attrs:
-            if feature_name not in fc_parameters:
-                fc_parameters[feature_name] = None
-            continue
-
-        # Create attributes dictionary
-        attrs = {}
-        for attr in feature_attrs:
-            attr_parts = attr.split('_')
-            attr_value = attr_parts[-1]
-            attr_name = '_'.join(attr_parts[:-1])
-            # Convert the attribute value to appropriate type
-            attrs[attr_name] = convert_value(attr_value)
-
-        # Handle multiple parameter combinations for same feature
-        if feature_name in fc_parameters:
-            if fc_parameters[feature_name] is None:
-                fc_parameters[feature_name] = [attrs]
-            elif isinstance(fc_parameters[feature_name], list):
-                fc_parameters[feature_name].append(attrs)
-            else:
-                fc_parameters[feature_name] = [fc_parameters[feature_name], attrs]
-        else:
-            fc_parameters[feature_name] = [attrs]
-
-    return fc_parameters
-
-def extract_tsfel_features(subset_features_df, COLUMN_CONFIG, n_tsfel_features):
-    """Extract TSFEL features from the top features."""
-    subset_features_df = subset_features_df.copy()
-    # Create and run TSFEL pipeline
-    pipeline = Pipeline([
-        ('imputation', ImputationWrapper(params=('interpolate', 'linear'), column_config=COLUMN_CONFIG)),
-        ('feature_extraction', TSFELLagFeatureExtractor(
-            n_lags=4,
-            domains=['statistical', 'temporal'],
-            column_config=COLUMN_CONFIG
-        )),
-        ('feature_selection', FeatureSelectionWrapper(strategy='correlation', column_config=COLUMN_CONFIG))
-    ])
-
-    # Transform data and get top TSFEL features
-    transformed_subset_df = pipeline.fit_transform(subset_features_df)
-    
-    clf = LGBMClassifier()
-    model = train_model(transformed_subset_df, transformed_subset_df[COLUMN_CONFIG['target_col']], 
-                       clf, COLUMN_CONFIG['protected_cols'])
-    tsfel_feature_names = get_top_features(model, transformed_subset_df, n_features=None, protected_cols=COLUMN_CONFIG['protected_cols'])
-    top_tsfel_features = select_tsfel_features(tsfel_feature_names, n_features=n_tsfel_features)
-    
-    # Generate TSFEL config file
-    tsfel_config_file = generate_tsfel_config(top_tsfel_features)
-    
-    return tsfel_config_file
-
-def extract_tsfresh_features(subset_features_df, COLUMN_CONFIG, n_tsfresh_features):
-    """Extract TSFresh features from the top features."""
-    subset_features_df = subset_features_df.copy()
-    # Create and run TSFresh pipeline
-    pipeline = Pipeline([
-        ('imputation', ImputationWrapper(params=('interpolate', 'linear'), column_config=COLUMN_CONFIG)),
-        ('feature_extraction', TSFreshLagFeatureExtractor(
-            n_lags=4,
-            default_fc_parameters=EfficientFCParameters(),
-            column_config=COLUMN_CONFIG,
-    )),
-        ('feature_selection', FeatureSelectionWrapper(strategy='correlation', column_config=COLUMN_CONFIG))
-    ])
-
-    # Transform data and get top TSfresh features
-    transformed_subset_df = pipeline.fit_transform(subset_features_df)
-        
-    clf = LGBMClassifier()
-
-    transformed_subset_df, column_mapping = clean_column_names(transformed_subset_df)
-
-    model = train_model(transformed_subset_df, transformed_subset_df[COLUMN_CONFIG['target_col']], 
-                       clf, COLUMN_CONFIG['protected_cols'])
-    
-    tsfresh_feature_names = get_top_features(model=model, X_train=transformed_subset_df, protected_cols=COLUMN_CONFIG['protected_cols'])
-    transformed_tsfresh_feature_names = restore_column_names(tsfresh_feature_names, column_mapping)
-    final_tsfresh_features = select_tsfresh_features(transformed_tsfresh_feature_names, n_features=n_tsfresh_features)
-
-    tsfresh_configuration = get_custumed_fc_parameters(final_tsfresh_features)
-
-    return tsfresh_configuration
-
+from extractor_preprocessing.feature_subset import get_subset_features_df, get_top_dataset_features
+from extractor_preprocessing.tsfel import extract_tsfel_features
+from extractor_preprocessing.tsfresh import extract_tsfresh_features
 
 
 def feature_extraction_preprocessing(data, COLUMN_CONFIG, n_features, n_tsfel_features, n_tsfresh_features):
