+++ extractor_preprocessing/tsfel.py
@@ -1,0 +1,103 @@
+import json
+from sklearn.pipeline import Pipeline
+import tsfel
+
+from transformers.feature_extractors import TSFELLagFeatureExtractor
+from transformers.feature_selection import FeatureSelectionWrapper
+from transformers.imputation import ImputationWrapper
+from lightgbm import LGBMClassifier
+
+from .common import train_model, get_top_features
+
+def generate_tsfel_config(selected_features=None):
+    """
+    Creates a modified copy of the TSFEL configuration file.
+    
+    Args:
+        selected_features (list): List of feature names to keep enabled in config
+    
+    This function:
+    1. Reads the original TSFEL features.json
+    2. Creates a new config with only selected features enabled
+    3. Saves it to tsfel_config.json
+    """
+    config_path = tsfel.__path__[0] + "/feature_extraction/features.json"
+    
+    try:
+        # Read original config
+        with open(config_path, 'r') as file:
+            tsfel_config = json.load(file)
+
+        # Create modified configuration
+        modified_config = {}
+        for domain in tsfel_config:
+            modified_config[domain] = {}
+            for feature in tsfel_config[domain]:
+                modified_config[domain][feature] = tsfel_config[domain][feature].copy()
+                # Disable all features by default
+                modified_config[domain][feature]["use"] = "no"
+                # Enable only selected features
+                if selected_features and feature in selected_features:
+                    modified_config[domain][feature]["use"] = "yes"
+
+        # Save modified configuration to new file
+        CONFIG_FILE_LOCATION = 'tsfel_config.json'
+        with open(CONFIG_FILE_LOCATION, 'w') as outfile:
+            json.dump(modified_config, outfile, indent=4)
+            
+        print("Successfully created modified tsfel_config.json")
+        
+    except FileNotFoundError:
+        print(f"Error: Could not find original config file at {config_path}")
+    except json.JSONDecodeError:
+        print("Error: Invalid JSON in configuration file")
+    except Exception as e:
+        print(f"Error occurred: {str(e)}")
+
+    return CONFIG_FILE_LOCATION
+
+# function to select from a list, the number elements that have '_0_' in its name
+def select_tsfel_features(features, n_features=10):
+    """
+    Selects features that have'_0_' in their name (ones generated by tsfel).
+    """
+    tsfel_features = [s for s in features if '_0_' in s]
+
+    #for each tsfel features split by _0_ and keep the second part
+    tsfel_feature_names = [s.split('_0_')[1] for s in tsfel_features]
+
+    tsfel_features = []
+    for feature_name in tsfel_feature_names:
+        if feature_name not in tsfel_features:
+            tsfel_features.append(feature_name)
+        if len(tsfel_features) == n_features:
+            break
+    return tsfel_features
+
+def extract_tsfel_features(subset_features_df, COLUMN_CONFIG, n_tsfel_features):
+    """Extract TSFEL features from the top features."""
+    subset_features_df = subset_features_df.copy()
+    # Create and run TSFEL pipeline
+    pipeline = Pipeline([
+        ('imputation', ImputationWrapper(params=('interpolate', 'linear'), column_config=COLUMN_CONFIG)),
+        ('feature_extraction', TSFELLagFeatureExtractor(
+            n_lags=4,
+            domains=['statistical', 'temporal'],
+            column_config=COLUMN_CONFIG
+        )),
+        ('feature_selection', FeatureSelectionWrapper(strategy='correlation', column_config=COLUMN_CONFIG))
+    ])
+
+    # Transform data and get top TSFEL features
+    transformed_subset_df = pipeline.fit_transform(subset_features_df)
+    
+    clf = LGBMClassifier()
+    model = train_model(transformed_subset_df, transformed_subset_df[COLUMN_CONFIG['target_col']], 
+                       clf, COLUMN_CONFIG['protected_cols'])
+    tsfel_feature_names = get_top_features(model, transformed_subset_df, n_features=None, protected_cols=COLUMN_CONFIG['protected_cols'])
+    top_tsfel_features = select_tsfel_features(tsfel_feature_names, n_features=n_tsfel_features)
+    
+    # Generate TSFEL config file
+    tsfel_config_file = generate_tsfel_config(top_tsfel_features)
+    
+    return tsfel_config_file
